incorporate ModalWrapper settings into Modal.sc.js
