+++ src/components/redirect_notification/modal_shared/ButtonContainer.js
@@ -1,4 +1,3 @@
-import React from "react";
 import * as s from "./Modal.sc";
 
 export default function ButtonContainer({ children, buttonCountNum }) {
