+++ src/components/redirect_notification/modal_shared/Modal.sc.js
@@ -1,9 +1,28 @@
-import { ModalWrapper as Wrapper } from "./ModalWrapper.sc"; //responsible for modal wrapper background, size and scaling
+import Box from "@mui/material/Box";
 import { zeeguuDarkOrange, zeeguuOrange } from "../../colors";
 import { OrangeRoundButton } from "../../allButtons.sc";
 import styled, { css } from "styled-components";
 
-const ModalWrapper = styled(Wrapper)`
+//responsible for modal wrapper background, size and scaling
+const ModalWrapper = styled(Box)`
+  position: absolute;
+  display: flex;
+  flex-direction: column;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 50%;
+  max-width: 500px;
+  background-color: white;
+  border: 0 !important;
+  border-radius: 0.65em;
+  padding: 32px 48px 32px 48px;
+  box-shadow:
+    0px 11px 15px -7px rgb(0 0 0 / 20%),
+    0px 24px 38px 3px rgb(0 0 0 / 14%),
+    0px 9px 46px 8px rgb(0 0 0 / 12%);
+  outline: none !important;
+
   p {
     line-height: 150%;
     text-align: left;
@@ -22,6 +41,16 @@ const ModalWrapper = styled(Wrapper)`
   .annotation {
     color: orange;
     font-weight: 500;
+  }
+
+  @media (max-width: 1200px) {
+    max-width: 500px;
+    width: 80%;
+  }
+
+  @media (max-width: 576px) {
+    padding: 24px 24px;
+    width: 80%;
   }
 `;
